+++ mayan/apps/dynamic_search/settings.py
@@ -2,6 +2,8 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.smart_settings.classes import SettingNamespace
 
+from .literals import DEFAULT_RESULTS_LIMIT
+
 namespace = SettingNamespace(label=_('Search'), name='search')
 
 setting_search_backend = namespace.add_setting(
@@ -21,4 +23,8 @@ setting_disable_simple_search = namespace.add_setting(
         'Disables the single term bar search leaving only the advanced '
         'search button.'
     )
+)
+setting_results_limit = namespace.add_setting(
+    global_name='SEARCH_RESULTS_LIMIT', default=DEFAULT_RESULTS_LIMIT,
+    help_text=_('Maximum number search results to fetch and display.')
 )